GetFilesSAS
I have developed a macro GetFilesFromAllDirNSubDirs to get the file path and file name of the files that you are looking for into a dataset.
In a real time, World, Business analysts keep their data in excel/CSV files. Sometimes BAs create folders for each year and subfolders within it for each month. This can go further down to regions, cities or clients and so on. Suddenly someone decides to put all this information into a dataset. It would be a real pain to search for these files which can be thousands or even more. Once you have file names and it location, we can write a program to loop thro’ all files to get the data into a dataset. 
Limitation to this program is user permissions to the directories and having access from SAS Server.